+++ tests/integration_test.py
@@ -182,7 +182,9 @@ class TestCreateContainerWithBinds(BaseTestCase):
             container = self.client.create_container(
                 'busybox',
                 ['ls', mount_dest], volumes={mount_dest: {}},
-                host_config=create_host_config(binds=binds)
+                host_config=create_host_config(
+                    binds=binds, network_mode='none'
+                )
             )
             container_id = container['Id']
             self.client.start(container_id)
@@ -222,7 +224,9 @@ class TestCreateContainerWithRoBinds(BaseTestCase):
             container = self.client.create_container(
                 'busybox',
                 ['ls', mount_dest], volumes={mount_dest: {}},
-                host_config=create_host_config(binds=binds)
+                host_config=create_host_config(
+                    binds=binds, network_mode='none'
+                )
             )
             container_id = container['Id']
             self.client.start(container_id)
@@ -243,6 +247,7 @@ class TestCreateContainerWithRoBinds(BaseTestCase):
         self.assertFalse(inspect_data['VolumesRW'][mount_dest])
 
 
+@unittest.skipIf(NOT_ON_HOST, 'Tests running inside a container; no syslog')
 class TestCreateContainerWithLogConfig(BaseTestCase):
     def runTest(self):
         config = docker.utils.LogConfig(
@@ -273,7 +278,9 @@ class TestCreateContainerReadOnlyFs(BaseTestCase):
     def runTest(self):
         ctnr = self.client.create_container(
             'busybox', ['mkdir', '/shrine'],
-            host_config=create_host_config(read_only=True)
+            host_config=create_host_config(
+                read_only=True, network_mode='none'
+            )
         )
         self.assertIn('Id', ctnr)
         self.tmp_containers.append(ctnr['Id'])
@@ -347,7 +354,9 @@ class TestStartContainerWithDictInsteadOfId(BaseTestCase):
 class TestCreateContainerPrivileged(BaseTestCase):
     def runTest(self):
         res = self.client.create_container(
-            'busybox', 'true', host_config=create_host_config(privileged=True)
+            'busybox', 'true', host_config=create_host_config(
+                privileged=True, network_mode='none'
+            )
         )
         self.assertIn('Id', res)
         self.tmp_containers.append(res['Id'])
@@ -591,7 +600,9 @@ class TestPort(BaseTestCase):
 
         container = self.client.create_container(
             'busybox', ['sleep', '60'], ports=list(port_bindings.keys()),
-            host_config=create_host_config(port_bindings=port_bindings)
+            host_config=create_host_config(
+                port_bindings=port_bindings, network_mode='bridge'
+            )
         )
         id = container['Id']
 
@@ -717,7 +728,9 @@ class TestCreateContainerWithVolumesFrom(BaseTestCase):
             )
         res2 = self.client.create_container(
             'busybox', 'cat', detach=True, stdin_open=True,
-            host_config=create_host_config(volumes_from=vol_names)
+            host_config=create_host_config(
+                volumes_from=vol_names, network_mode='none'
+            )
         )
         container3_id = res2['Id']
         self.tmp_containers.append(container3_id)
@@ -760,7 +773,8 @@ class TestCreateContainerWithLinks(BaseTestCase):
 
         res2 = self.client.create_container(
             'busybox', 'env', host_config=create_host_config(
-                links={link_path1: link_alias1, link_path2: link_alias2}
+                links={link_path1: link_alias1, link_path2: link_alias2},
+                network_mode='none'
             )
         )
         container3_id = res2['Id']
@@ -781,7 +795,8 @@ class TestRestartingContainer(BaseTestCase):
     def runTest(self):
         container = self.client.create_container(
             'busybox', ['sleep', '2'], host_config=create_host_config(
-                restart_policy={"Name": "always", "MaximumRetryCount": 0}
+                restart_policy={"Name": "always", "MaximumRetryCount": 0},
+                network_mode='none'
             )
         )
         id = container['Id']
@@ -873,8 +888,8 @@ class TestRunContainerStreaming(BaseTestCase):
         id = container['Id']
         self.client.start(id)
         self.tmp_containers.append(id)
-        socket = self.client.attach_socket(container, ws=False)
-        self.assertTrue(socket.fileno() > -1)
+        sock = self.client.attach_socket(container, ws=False)
+        self.assertTrue(sock.fileno() > -1)
 
 
 class TestPauseUnpauseContainer(BaseTestCase):
@@ -910,7 +925,7 @@ class TestCreateContainerWithHostPidMode(BaseTestCase):
     def runTest(self):
         ctnr = self.client.create_container(
             'busybox', 'true', host_config=create_host_config(
-                pid_mode='host'
+                pid_mode='host', network_mode='none'
             )
         )
         self.assertIn('Id', ctnr)
@@ -945,7 +960,7 @@ class TestRemoveLink(BaseTestCase):
 
         container2 = self.client.create_container(
             'busybox', 'cat', host_config=create_host_config(
-                links={link_path: link_alias}
+                links={link_path: link_alias}, network_mode='none'
             )
         )
         container2_id = container2['Id']
@@ -1376,8 +1391,8 @@ class TestLoadConfig(BaseTestCase):
         f.write('email = sakuya@scarlet.net')
         f.close()
         cfg = docker.auth.load_config(cfg_path)
-        self.assertNotEqual(cfg[docker.auth.INDEX_URL], None)
-        cfg = cfg[docker.auth.INDEX_URL]
+        self.assertNotEqual(cfg[docker.auth.INDEX_NAME], None)
+        cfg = cfg[docker.auth.INDEX_NAME]
         self.assertEqual(cfg['username'], 'sakuya')
         self.assertEqual(cfg['password'], 'izayoi')
         self.assertEqual(cfg['email'], 'sakuya@scarlet.net')
@@ -1406,7 +1421,7 @@ class TestLoadJSONConfig(BaseTestCase):
 
 class TestAutoDetectVersion(unittest.TestCase):
     def test_client_init(self):
-        client = docker.Client(version='auto')
+        client = docker.Client(base_url=DEFAULT_BASE_URL, version='auto')
         client_version = client._version
         api_version = client.version(api_version=False)['ApiVersion']
         self.assertEqual(client_version, api_version)
@@ -1415,7 +1430,7 @@ class TestAutoDetectVersion(unittest.TestCase):
         client.close()
 
     def test_auto_client(self):
-        client = docker.AutoVersionClient()
+        client = docker.AutoVersionClient(base_url=DEFAULT_BASE_URL)
         client_version = client._version
         api_version = client.version(api_version=False)['ApiVersion']
         self.assertEqual(client_version, api_version)
@@ -1423,7 +1438,7 @@ class TestAutoDetectVersion(unittest.TestCase):
         self.assertEqual(client_version, api_version_2)
         client.close()
         with self.assertRaises(docker.errors.DockerException):
-            docker.AutoVersionClient(version='1.11')
+            docker.AutoVersionClient(base_url=DEFAULT_BASE_URL, version='1.11')
 
 
 class TestConnectionTimeout(unittest.TestCase):
@@ -1484,8 +1499,18 @@ class TestRegressions(BaseTestCase):
         self.client.start(
             self.client.create_container('busybox', ['true'])
         )
-        result = self.client.containers(trunc=True)
+        result = self.client.containers(all=True, trunc=True)
         self.assertEqual(len(result[0]['Id']), 12)
+
+    def test_647(self):
+        with self.assertRaises(docker.errors.APIError):
+            self.client.inspect_image('gensokyo.jp//kirisame')
+
+    def test_649(self):
+        self.client.timeout = None
+        ctnr = self.client.create_container('busybox', ['sleep', '2'])
+        self.client.start(ctnr)
+        self.client.stop(ctnr)
 
 
 if __name__ == '__main__':